+++ src/Modules/StoreUser/StoreUserControl.php
@@ -159,11 +159,9 @@ class StoreUserControl extends Control
 				/* team status */
 				if ($this->storePermissions->mayEditStore($storeId)) {
 					$this->pageHelper->addContent(
-						$this->v_utils->v_field(
-							$this->view->u_legacyStoreTeamStatus($store),
-							$this->translator->trans('status'),
-							['class' => 'ui-padding']
-						),
+						$this->view->vueComponent('vue-store-teamstatus', 'StoreTeamStatus', [
+							'storeId' => $storeId,
+						]),
 						CNT_LEFT
 					);
 				}